Updated AdvancedFiltering, SignUp, and ThemeToggle components; refactored global CSS; and modified layout.js to remove metadata and viewport exports
